#5-create-user-model
Testing user updae

+++ payment-system/tests/test_user.py
@@ -16,6 +16,7 @@ class TestUser(TestCase):
                     phone_number='+55 21 99999-999',
                     mail_address='test@test_users.com',
                     password=self.pwd)
+        self.test.save()
 
     def test_hashed_password(self):
         """
@@ -35,8 +36,6 @@ class TestUser(TestCase):
         used username
         """
 
-        self.test.save()
-
         user2 = User(name="Testing User 02",
                      username="test01",
                      adress="0, Dummy Street, 219875-456",
@@ -49,40 +48,45 @@ class TestUser(TestCase):
         result = len(User.nodes.filter(username=self.test.username))
         self.assertEqual(1, result)
 
-        self.test.delete()
-
     def test_register_user_with_used_username_return_none(self):
         """
         Should get none when saving user with an already
         used username
         """
-
-        user1 = User(name="Testing User 02",
-                     username="test02",
-                     password='weak password')
-
-        result1 = user1.save()
-
-        self.assertIsNotNone(result1, msg="Get none on first save")
-
         user2 = User(name="Testing User 02",
-                     username="test02",
+                     username="test01",
                      password='weak password')
 
         result2 = user2.save()
 
         self.assertIsNone(result2, msg="Item was saved with same username")
 
-        user1.delete()
-
     def test_update_user(self):
         """
         Update user means save an user with same username,
         but once is an uptade, not a register, should proceed
         """
+        # Load the already created user
+        user1 = User.nodes.get_or_none(username='test01')
+
+        # update mail address:
+        user1.mail_address = 'user1@mail_adress.com'
+
+        # Try saving:
+        result = user1.save()
+
+        # Test if result is None (it is, if saving failed)
+        self.assertIsNotNone(result)
+        # Load again and test if e-mail was changed
+        user2 = User.nodes.get_or_none(username='test01')
+        self.assertEqual(user2.mail_address, 'user1@mail_adress.com')
 
     def test_login_correct_password(self):
         pass
 
     def test_login_incorrect_password(self):
         pass
+
+    def tearDown(self):
+        if self.test.save() is not None:
+            self.test.delete()
